salary datatype change

+++ src/main/java/com/ust/docker_jenkin_springboot/model/Employee.java
@@ -15,6 +15,6 @@ public class Employee {
     private Long id;
     private String name;
     private String position;
-    private double salary;
+    private int salary;
     private String company;
 }
